.gitlab/python/requirements.txt
-i https://pypi.org/simple
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
diff-match-patch==20230430; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
python-gitlab==4.4.0; python_full_version >= '3.8.0'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==2.2.1; python_version >= '3.8'
